tribunal/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tribunal"
version = "2.0.0a1"
description = "Quality gates for AI-generated code — TDD enforcement, secret scanning, and audit trails for coding agents."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "thebotclub" }]
keywords = ["ai", "quality-gates", "tdd", "secret-scanning", "coding-agent", "ci", "hooks"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
]

[project.urls]
Homepage = "https://tribunal.dev"
Repository = "https://github.com/thebotclub/tribunal"
Documentation = "https://github.com/thebotclub/tribunal/blob/main/README.md"

[project.scripts]
tribunal = "tribunal.cli:main"
tribunal-gate = "tribunal.gate:main"

[tool.hatch.build.targets.wheel]
packages = ["src/tribunal"]
artifacts = ["src/tribunal/py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.sdist]
include = ["src/tribunal"]
